Always create the daemon user
Currently the daemon user (`phd`) is created in `Class[phabricator::daemons]`. It seems, however, that this user can be required on hosts that _aren't_ running the actual daemons though. For example, a host that is proxying VCS traffic requires an Almanac device and associated private key, which must be owned by `phd.user`. From the source code:

```
$tmp = new TempFile();
list($err) = exec_manual('chown %s %s', $phd_user, $tmp);
if ($err) {
  throw new PhutilArgumentUsageException(
    pht(
      'Unable to change ownership of an identity file to daemon user '.
      '"%s". Run this command as %s or root.',
      $phd_user,
      $phd_user));
}
```

Furthermore, it seems necessary that the Almanac key is group-readable as well (otherwise the PHP-FPM user would not be able to read this file).

I guess we _could_ try to be a bit more clever and only create the user if `Class[phabricator::almanac]` or `Class[phabricator::daemons]` are defined, but this seems overly complicated. Practically speaking, I think it's at least somewhat reasonable to just _always_ create the daemon user, even if it is possibly unnecessary for the given node.